[REF] Package restructuring for better growth
