+++ pyls/matlab/__init__.py
@@ -1,0 +1,4 @@
+__all__ = ['import_matlab_result']
+
+
+from .io import import_matlab_result
