+++ pyls/types/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+__all__ = ['behavioral_pls', 'meancentered_pls']
+
+from .behavioral import behavioral_pls
+from .meancentered import meancentered_pls
